+++ src/tests/simulation/tests/integration_tests/main.rs
@@ -22,7 +22,7 @@
 #![feature(lazy_cell)]
 #![feature(drain_filter)]
 
-mod backfill;
+mod backfill_tests;
 mod batch;
 mod recovery;
 mod scale;
